Move contract methods to concerns
